Fixed: Gold VIP Donation with URL grants both VIP gold and a song request
