+++ server/src/services/rewardService.ts
@@ -27,8 +27,24 @@ export default class RewardService {
         const user = await this.getUserForEvent(donation.from);
 
         this.addUserPoints(user, donation);
-        this.addSongsToQueue(donation);
-        this.addGoldStatus(user, donation);
+
+        if (await this.addGoldStatus(user, donation) && user) {
+            // Use goldsong for queue if possible
+            const matches = this.getSongsForQueue(donation.message);
+            for (const match of matches) {
+                try {
+                    await this.songService.addGoldSong(match, user);
+
+                    // Only accept one song per donation.
+                    break;
+                } catch {
+                    // Ignore any invalid songs
+                    // Maybe consider URLs from unknown services in the future.
+                }
+            }
+        } else {
+            this.addSongsToQueue(donation);
+        }
     }
 
     public async processBits(bits: IBitsMessage) {
@@ -74,7 +90,7 @@ export default class RewardService {
             this.addSubUserPoints(giftingUser, giftedMonths, PointLogType.GiftSubGiver, false);
 
             if (plan === SubscriptionPlan.Tier3) {
-                // Both gifter and receiver gets halb the amount of VIP gold.
+                // Both gifter and receiver gets half the amount of VIP gold.
                 // We assume that the user on the receiving end will be covered by a streamlabs event.
                 if (giftingUser) {
                     this.userService.addVipGoldWeeks(giftingUser, giftedMonths, "Gifted T3 sub");
@@ -96,14 +112,17 @@ export default class RewardService {
         return user;
     }
 
-    private async addGoldStatus(user: IUser | undefined, donation: IDonationMessage) {
+    private async addGoldStatus(user: IUser | undefined, donation: IDonationMessage): Promise<boolean> {
         const amountPerMonth = parseFloat(await this.settings.getValue(BotSettings.GoldStatusDonationAmount));
         const goldMonths = Math.floor(donation.amount / amountPerMonth);
         if (goldMonths > 0) {
             if (user) {
-                this.userService.addVipGoldWeeks(user, goldMonths * 4, "Donation");
+                await this.userService.addVipGoldWeeks(user, goldMonths * 4, "Donation");
+                return true;
             }
         }
+
+        return false;
     }
 
     private async addUserPoints(user: IUser | undefined, donation: IDonationMessage) {
@@ -128,23 +147,25 @@ export default class RewardService {
         }
     }
 
-    private async addSongsToQueue(donation: IDonationMessage) {
+    private getSongsForQueue(message: string): string[] {
         const urlRegex: RegExp = /(https?:\/\/[^\s]+)/g;
+        return urlRegex.exec(message) ?? [];
+    }
+
+    private async addSongsToQueue(donation: IDonationMessage) {
         const amountRequired = parseFloat(await this.settings.getValue(BotSettings.SongRequestDonationAmount));
 
         if (donation.amount >= amountRequired) {
-            const matches = urlRegex.exec(donation.message);
-            if (matches) {
-                for (const match of matches) {
-                    try {
-                        await this.songService.addSong(match, RequestSource.Donation, donation.from);
+            const matches = this.getSongsForQueue(donation.message);
+            for (const match of matches) {
+                try {
+                    await this.songService.addSong(match, RequestSource.Donation, donation.from);
 
-                        // Only accept one song per donation.
-                        return;
-                    } catch {
-                        // Ignore any invalid songs
-                        // Maybe consider URLs from unknown services in the future.
-                    }
+                    // Only accept one song per donation.
+                    return;
+                } catch {
+                    // Ignore any invalid songs
+                    // Maybe consider URLs from unknown services in the future.
                 }
             }
         }
